requirements.txt
ipykernel
matplotlib
numpy
pandas
pyyaml